pyproject.toml
[project]
name = "llm-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Prathamesh-Ghatole", email = "prathamesh.s.ghatole@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "pillow>=11.3.0",
    "polars>=1.33.0",
    "pypdfium2>=4.30.0",
]

[project.scripts]
llm-extractor = "llm_extractor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
